requirements.txt
absl-py==0.12.0
apache-beam==2.29.0
astunparse==1.6.3
attrs==20.3.0
avro-python3==1.9.2.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
camelot-py==0.8.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
decorator==5.0.7
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
et-xmlfile==1.1.0
ExtractTable==2.2.0
fastavro==1.4.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.15.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.2
imutils==0.5.4
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.1
matplotlib-inline==0.1.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
object-detection @ file:///C:/Users/azzub/OneDrive/Desktop/RND/Local/myOCR_bank_statement/train/models/research
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
pdf2image==1.14.0
pdfminer.six==20201018
pickleshare==0.7.5
pikepdf==2.12.2
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pymongo==3.11.3
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-slugify==5.0.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
sortedcontainers==2.4.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==0.30.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tqdm
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
XlsxWriter
zipp==3.4.1
